Cross browser: extract module path from the importer body
This feature is, just like the original approach, a hack that is
quite fragile. We inspect the importer function and parse the
string from the "import('../my-module')" call.

This works under the assumption that:
1. Vite or whatever bundler has put the FULL PATH from the root
into that import call when transpiling.
2. The importer function is relativly simple and will not throw
off our simple regex match.

In any case, the old Chromium only approach is still available
under the hood as an optional strategy that can be selected
at runtime. The tests show both to be working.