Allow using custom strategies expressed as functions (#16)
This PR contributes towards the idea discussed in #13 - new strategy which would
allow the user to define a "hint" for `dynamicImportWithRetry` to use un case
a retry is needed.

Instead of introducing the property (symbol) that I proposed in
https://github.com/fatso83/retry-dynamic-import/issues/13#issuecomment-3239576867
I instead introduce ability to pass a **custom strategy**, which allows to build
a feature like I suggested on ones own, but without enforcing the way it will be
built

Implementation which I suggested in the comment is now added as a test.

With the ability to pass strategies from outside the overall solution may be
extended in a more tree-shakeable way. Current two strategies are both bundled
regardless of which is used..